v.0.8.12 Decoder/Reader refactored + reduces memory usage.

+++ engine/include/analyzer/report/label.h
@@ -11,10 +11,10 @@
 
 typedef struct vmcu_label {
 
-    uint16_t id;
-    uint16_t addr;
+    uint32_t id;
+    uint32_t addr;
 
-    int32_t n_xfrom;
+    uint32_t n_xfrom;
     vmcu_xref_t *xfrom;
 
 } vmcu_label_t;
