+++ engine/include/analyzer/util/util.h
@@ -3,8 +3,11 @@
 #ifndef VMCU_UTIL_H
 #define VMCU_UTIL_H
 
+// C Headers
+#include <inttypes.h>
+
 typedef struct vmcu_instr vmcu_instr_t;
 
-extern int vmcu_resolve_flow(const vmcu_instr_t *instr);
+extern int64_t vmcu_resolve_flow(const vmcu_instr_t *instr);
 
 #endif
