+++ engine/include/misc/stringmanip.h
@@ -3,8 +3,11 @@
 #ifndef VMCU_STRINGMANIP_H
 #define VMCU_STRINGMANIP_H
 
+// C Headers
+#include <inttypes.h>
+
 extern int vmcu_strpos(const char *str, const char *sub);
 extern char* vmcu_substr(const char *str, const int from, const int to);
-extern int vmcu_htoi(const char *input);
+extern int64_t vmcu_htoi(const char *input);
 
 #endif
