+++ engine/src/analyzer/util/util.c
@@ -4,16 +4,18 @@
 #include "engine/include/analyzer/util/util.h"
 #include "engine/include/analyzer/report/instr.h"
 
-int vmcu_resolve_flow(const vmcu_instr_t *instr) {
+int64_t vmcu_resolve_flow(const vmcu_instr_t *instr) {
 
-    /* this function only accepts branches, calls/jumps */
+    /* this function only accepts branches, calls/jumps
+     * todo: handle flash wrap arounds
+     * */
 
     if(instr->key == VMCU_IKEY_CALL || instr->key == VMCU_IKEY_JMP)
         return instr->src.p;
 
     if(instr->src.s < 0) {
 
-        if((instr->src.s * -1) > instr->addr)
+        if((uint32_t) (instr->src.s * -1) > instr->addr)
             return -1;
     }
