+++ engine/src/misc/bitmanip.c
@@ -3,14 +3,14 @@
 // Project Headers (engine)
 #include "engine/include/misc/bitmanip.h"
 
-int vmcu_extr(const int opcode, int from, int to, int offs) {
+uint32_t vmcu_extr(const uint32_t opcode, uint8_t from, uint8_t to, uint8_t offs) {
 
-    int res = 0;
+    uint32_t res = 0;
 
-    for(int i = from; i < to; i++) {
+    for(uint8_t i = from; i < to; i++) {
 
-        const int bit = (((0x01 << i) & opcode) >> i);
-        res += (bit << (i - from) + offs);
+        const uint8_t bit = (((0x01 << i) & opcode) >> i);
+        res += (bit << ((i - from) + offs));
     }
 
     return res;
